refactor(ui): move theme to dbflux_components, shim in dbflux_ui (S1) (#102)
