+++ crates/dbflux_components/src/lib.rs
@@ -22,6 +22,7 @@ pub mod primitives;
 pub mod result_panel;
 pub mod result_view;
 pub mod saved_chart;
+pub mod theme;
 
 pub use saved_chart::{
     SavedChart, SavedChartManager, SavedChartRefreshPolicy, SavedChartStore, TimeRangePreset,
